Initial functional backend: orders, payments and order closing

+++ GersoPersonalizados.Api/Dtos/CreatePaymentDto.cs
@@ -2,10 +2,11 @@
 
 public class CreatePaymentDto
 {
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }          
     public decimal Amount { get; set; }
+
     public string Method { get; set; } = "CASH";   // NEQUI, DAVIPLATA, TRANSFER, CASH
     public string? Reference { get; set; }
     public string? Notes { get; set; }
-    public DateTime? PaidAt { get; set; } // si no lo mandas, usamos ahora
+    public DateTime? PaidAt { get; set; }
 }
